refactor(frontend): extract into common component
