+++ apps/frontend/app/routes/_dashboard.media.$action.$lot.tsx
@@ -9,7 +9,6 @@ import {
 	Group,
 	Loader,
 	Menu,
-	MultiSelect,
 	Pagination,
 	Select,
 	Stack,
@@ -56,6 +55,7 @@ import { z } from "zod";
 import { zx } from "zodix";
 import {
 	ApplicationGrid,
+	CollectionsFilter,
 	DebouncedSearchInput,
 	FiltersModal,
 } from "~/components/common";
@@ -65,7 +65,6 @@ import {
 	useAppSearchParam,
 	useApplicationEvents,
 	useCoreDetails,
-	useUserCollections,
 	useUserDetails,
 	useUserPreferences,
 } from "~/lib/hooks";
@@ -136,7 +135,7 @@ export const loader = unstable_defineLoader(async ({ request, params }) => {
 				generalFilter: z
 					.nativeEnum(MediaGeneralFilter)
 					.default(defaultFilters.mineGeneralFilter),
-				collection: commaDelimitedString,
+				collections: commaDelimitedString,
 				invertCollection: zx.BoolAsString.optional(),
 			});
 			const { metadataList } = await serverGqlService.authenticatedRequest(
@@ -149,7 +148,7 @@ export const loader = unstable_defineLoader(async ({ request, params }) => {
 						sort: { order: urlParse.sortOrder, by: urlParse.sortBy },
 						filter: {
 							general: urlParse.generalFilter,
-							collection: urlParse.collection,
+							collections: urlParse.collections,
 						},
 						invertCollection: urlParse.invertCollection,
 					},
@@ -222,7 +221,7 @@ export default function Page() {
 			defaultFilters.mineGeneralFilter ||
 		loaderData.mediaList?.url.sortOrder !== defaultFilters.mineSortOrder ||
 		loaderData.mediaList?.url.sortBy !== defaultFilters.mineSortBy ||
-		loaderData.mediaList?.url.collection !== defaultFilters.mineCollection;
+		loaderData.mediaList?.url.collections !== defaultFilters.mineCollection;
 
 	return (
 		<Container>
@@ -525,7 +524,6 @@ const MediaSearchItem = (props: {
 
 const FiltersModalForm = () => {
 	const loaderData = useLoaderData<typeof loader>();
-	const collections = useUserCollections();
 	const [_, { setP }] = useAppSearchParam(loaderData.cookieName);
 
 	if (!loaderData.mediaList) return null;
@@ -579,21 +577,9 @@ const FiltersModalForm = () => {
 				</ActionIcon>
 			</Flex>
 			<Flex gap="xs" align="center">
-				<MultiSelect
-					placeholder="Select a collection"
-					defaultValue={loaderData.mediaList.url.collection}
-					data={[
-						{
-							group: "My collections",
-							items: collections.map((c) => ({
-								value: c.id.toString(),
-								label: c.name,
-							})),
-						},
-					]}
-					onChange={(v) => setP("collection", v.join(","))}
-					clearable
-					searchable
+				<CollectionsFilter
+					cookieName={loaderData.cookieName}
+					collections={loaderData.mediaList.url.collections}
 				/>
 				<Checkbox
 					label="Invert"
