+++ apps/frontend/app/routes/_dashboard.media.people.$action.tsx
@@ -7,7 +7,6 @@ import {
 	Flex,
 	Group,
 	Loader,
-	MultiSelect,
 	Pagination,
 	Select,
 	Stack,
@@ -45,16 +44,13 @@ import { z } from "zod";
 import { zx } from "zodix";
 import {
 	ApplicationGrid,
+	CollectionsFilter,
 	DebouncedSearchInput,
 	FiltersModal,
 } from "~/components/common";
 import { BaseMediaDisplayItem, PersonDisplayItem } from "~/components/media";
 import { commaDelimitedString } from "~/lib/generals";
-import {
-	useAppSearchParam,
-	useCoreDetails,
-	useUserCollections,
-} from "~/lib/hooks";
+import { useAppSearchParam, useCoreDetails } from "~/lib/hooks";
 import {
 	getEnhancedCookieName,
 	redirectUsingEnhancedCookieSearchParams,
@@ -98,7 +94,7 @@ export const loader = unstable_defineLoader(async ({ request, params }) => {
 			const urlParse = zx.parseQuery(request, {
 				sortBy: z.nativeEnum(PersonSortBy).default(defaultFilters.sortBy),
 				orderBy: z.nativeEnum(GraphqlSortOrder).default(defaultFilters.orderBy),
-				collection: commaDelimitedString,
+				collections: commaDelimitedString,
 				invertCollection: zx.BoolAsString.optional(),
 			});
 			const { peopleList } = await serverGqlService.authenticatedRequest(
@@ -108,7 +104,7 @@ export const loader = unstable_defineLoader(async ({ request, params }) => {
 					input: {
 						search: { page, query },
 						sort: { by: urlParse.sortBy, order: urlParse.orderBy },
-						filter: { collection: urlParse.collection },
+						filter: { collections: urlParse.collections },
 						invertCollection: urlParse.invertCollection,
 					},
 				},
@@ -370,7 +366,6 @@ const commitPerson = async (
 
 const FiltersModalForm = () => {
 	const loaderData = useLoaderData<typeof loader>();
-	const collections = useUserCollections();
 	const [_, { setP }] = useAppSearchParam(loaderData.cookieName);
 
 	if (!loaderData.peopleList) return null;
@@ -402,21 +397,9 @@ const FiltersModalForm = () => {
 				</ActionIcon>
 			</Flex>
 			<Flex gap="xs" align="center">
-				<MultiSelect
-					placeholder="Select a collection"
-					defaultValue={loaderData.peopleList.url.collection}
-					data={[
-						{
-							group: "My collections",
-							items: collections.map((c) => ({
-								value: c.id.toString(),
-								label: c.name,
-							})),
-						},
-					]}
-					onChange={(v) => setP("collection", v.join(","))}
-					clearable
-					searchable
+				<CollectionsFilter
+					cookieName={loaderData.cookieName}
+					collections={loaderData.peopleList.url.collections}
 				/>
 				<Checkbox
 					label="Invert"
